Made a number of changes to MAMClient that enabled successful building on both 64bit and 32bit machines. In addition to this, the work performed also enables successful login of large and small usernames, corrected some memory issues, and resolved some code completeness issues.

+++ cpp/MAMClient/Utilities.cpp
@@ -20,7 +20,7 @@ unsigned char rc5_cipher[] = {
     0x8D, 0xF9, 0x16, 0x64, 0x5B, 0xAB, 0x42, 0x66
 };
 
-unsigned long long server_start_time = 0;
+unsigned long server_start_time = 0;
 
 unsigned long timeGetTime( void )
 {
